Fin Backend

+++ quiz-api/controller.py
@@ -2,23 +2,29 @@ import json
 from multiprocessing.connection import answer_challenge
 import sqlite3
 import pathlib
+from turtle import position
 from Question import Question
 from Answer import Answer
 from Participation import Participation
 from ParticipationAnswer import ParticipationAnswer
 
-
 def ConvertJSONToQuestion(data):
 	q = Question(data["position"], data["title"], data["text"], data["image"])
 	return q
 
-def AddQuestion(input):
-    question = ConvertJSONToQuestion(input)
+def GetQuestionIdFromPosition(db_connection, position):
+    get_question = db_connection.execute(
+        "SELECT Id FROM Question WHERE Position = ?",
+        (position,))
 
-    answers_list = list()
-    for answer in input['possibleAnswers']:
-        answers_list.append((answer["text"], question.position, answer["isCorrect"]))
+    question = get_question.fetchall()
+    
+    if len(question) == 0:
+        return -1
 
+    return question[0][0]
+
+def QuizInfo():
     #création d'un objet connection
     path = pathlib.Path(__file__).parent.resolve()/"Database.db"
 
@@ -30,10 +36,53 @@ def AddQuestion(input):
     # start transaction
     db_connection.execute("begin")
 
+    get_question_count = db_connection.execute(
+        "SELECT COUNT(*) FROM Question;",)
+
+    size = get_question_count.fetchall()[0][0]
+
+    get_player_scores = db_connection.execute(
+        "SELECT Player, Score FROM Participation ORDER BY Score DESC;",)
+
+    player_scores = get_player_scores.fetchall()
+
+    scores = list()
+    for s in player_scores:
+        scores.append({"playerName": s[0], "score": s[1]})
+
+    db_connection.execute("commit")
+    
+    return {"size": size, "scores": scores}
+
+def AddQuestion(input):
+    #création d'un objet connection
+    path = pathlib.Path(__file__).parent.resolve()/"Database.db"
+
+    db_connection = sqlite3.connect(path)
+    # set the sqlite connection in "manual transaction mode"
+    # (by default, all execute calls are performed in their own transactions, not what we want)
+    db_connection.isolation_level = None
+    
+    # start transaction
+    db_connection.execute("begin")
+
+    question = ConvertJSONToQuestion(input)
+
+    # increment position for higher position questions
+    update_questions = db_connection.execute(
+        "UPDATE Question SET Position = Position + ? WHERE Position >= ?",
+        (1, question.position))
+
     # save the question to db
     insert_question = db_connection.execute(
         "INSERT INTO Question (Position, Title, Text, Image) VALUES (?, ?, ?, ?)",
         (question.position, question.title, question.text, question.image))
+
+    question_id = GetQuestionIdFromPosition(db_connection, question.position)
+
+    answers_list = list()
+    for answer in input['possibleAnswers']:
+        answers_list.append((answer["text"], question_id, answer["isCorrect"]))
 
     # save the answer to db
     insert_answers = db_connection.executemany(
@@ -41,6 +90,8 @@ def AddQuestion(input):
     
     #send the request
     db_connection.execute("commit")
+
+    return '', 200
 
 def FetchQuestion(position):
     #création d'un objet connection
@@ -56,22 +107,26 @@ def FetchQuestion(position):
 
     # save the question to db
     get_question = db_connection.execute(
-        "SELECT Position, Title, Text, Image FROM Question WHERE Position = ?",
-        position)
+        "SELECT Id, Position, Title, Text, Image FROM Question WHERE Position = ?",
+        (position,))
     
     question = get_question.fetchall()
-    q = Question(question[0][0], question[0][1], question[0][2], question[0][3])
+    
+    if len(question) == 0:
+        return '', 404
+    
+    q = Question(question[0][1], question[0][2], question[0][3], question[0][4])
 
     get_answers = db_connection.execute(
         "SELECT Text, Id, Is_Correct FROM Answer WHERE Question_Id = ?",
-        position)
+        (question[0][0],))
     
     answers = get_answers.fetchall()
 
     answers_list = list()
     for answer in answers:
-        a = Answer(answer[0], answer[1], answer[2])
-        answers_list.append(json.loads(a.ConvertToJson()))
+        a = Answer(answer[0], answer[1], True if answer[2] == 1 else False)
+        answers_list.append(a.ConvertToJson())
 
     return_json = json.loads(q.ConvertToJson())
     return_json["possibleAnswers"] = answers_list
@@ -79,15 +134,11 @@ def FetchQuestion(position):
     #send the request
     db_connection.execute("commit")
 
-    return return_json
+    return return_json, 200
 
-def UpdateQuestion(input):
+def UpdateQuestion(old_position, input):
     #création d'un objet connection
     path = pathlib.Path(__file__).parent.resolve()/"Database.db"
-
-    answers_list = list()
-    for answer in input['answers']:
-        answers_list.append((answer["text"], input["position"], answer["isCorrect"]))
     
     db_connection = sqlite3.connect(path)
     # set the sqlite connection in "manual transaction mode"
@@ -97,15 +148,46 @@ def UpdateQuestion(input):
     # start transaction
     db_connection.execute("begin")
     
+    question_id = GetQuestionIdFromPosition(db_connection, old_position)
+
+    if question_id == -1:
+        return '', 404
+    
     # delete old answers
     delete_answers = db_connection.execute(
         "DELETE FROM Answer WHERE Question_Id = ?",
-        (input["position"],))
+        (question_id,))
 
-    # update question
-    update_question = db_connection.execute(
-        "UPDATE Question SET Position = ?, Title = ?, Text = ?, Image = ? WHERE Position = ?",
-        (input["position"], input["title"], input["text"], input["image"], input["position"]))
+    answers_list = list()
+    for answer in input['possibleAnswers']:
+        answers_list.append((answer["text"], question_id, answer["isCorrect"]))
+
+    old_position = int(old_position)
+    
+    switch_question_positions = db_connection.execute(
+            "UPDATE Question SET Position = ?, Title = ?, Text = ?, Image = ? WHERE Position = ? AND Id = ?",
+            (input["position"], input["title"], input["text"], input["image"], old_position, question_id))
+
+    if input["position"] == old_position:
+        print("same")
+        # update question
+        update_question = db_connection.execute(
+            "UPDATE Question SET Position = ?, Title = ?, Text = ?, Image = ? WHERE Position = ?",
+            (input["position"], input["title"], input["text"], input["image"], old_position))
+    
+    if input["position"] < old_position:
+        print("lesser than")
+        # increment other questions position
+        update_question_positions = db_connection.execute(
+            "UPDATE Question SET Position = Position + ? WHERE Position >= ? AND Position < ? AND NOT Id = ?",
+            (1, input["position"], old_position, question_id))
+    
+    if input["position"] > old_position:
+        print("greater than")
+        # decrement other questions position
+        update_question_positions = db_connection.execute(
+            "UPDATE Question SET Position = Position - ? WHERE Position <= ? AND Position > ? AND NOT Id = ?",
+            (1, input["position"], old_position, question_id))
     
     # save new answers
     insert_answers = db_connection.executemany(
@@ -114,7 +196,7 @@ def UpdateQuestion(input):
     #send the request
     db_connection.execute("commit")
 
-    return
+    return '', 200
 
 def RemoveQuestion(position):
     #création d'un objet connection
@@ -128,19 +210,34 @@ def RemoveQuestion(position):
     # start transaction
     db_connection.execute("begin")
 
+    question_id = GetQuestionIdFromPosition(db_connection, position)
+
+    if question_id == -1:
+        return '', 404
+    
     delete_answers = db_connection.execute(
         "DELETE FROM Answer WHERE Question_Id = ?",
-        position)
+        (question_id,))
 
     delete_Question = db_connection.execute(
         "DELETE FROM Question WHERE Position = ?",
-        position)
+        (position,))
+
+    # decrement position for higher position questions
+    update_questions = db_connection.execute(
+        "UPDATE Question SET Position = Position - ? WHERE Position > ?",
+        (1, position))
 
     #send the request
     db_connection.execute("commit")
 
+    return '', 204
+
 def AddParticipation(input):
-    print(input)
+    quiz_info = QuizInfo()
+    if len(input["answers"]) != quiz_info["size"]:
+        return '', 400
+
     p = Participation(input["playerName"])
     
     #création d'un objet connection
@@ -165,25 +262,31 @@ def AddParticipation(input):
     participation_id = get_participation.fetchall()[0][0]
 
     participation_answers_list = list()
-    question_number = 0
+    question_position = 0
     score = 0
     for answer in input["answers"]:
-        question_number += 1
+        question_position += 1
+        question_id = GetQuestionIdFromPosition(db_connection, question_position)
         get_answer = db_connection.execute(
             "SELECT Id, Is_Correct FROM Answer WHERE Question_Id = ?",
-            (question_number,))
+            (question_id,))
         fetch_answer = get_answer.fetchall()
-        if(fetch_answer[answer-1][1] == '1'):
-            score += 1
+        score += fetch_answer[answer-1][1]
         participation_answers_list.append((participation_id, fetch_answer[answer-1][0]))
     
     insert_answers = db_connection.executemany(
         "INSERT INTO Participation_Answer(Participation_Id, Answer_Id) VALUES(?,?)", participation_answers_list)
 
+    update_score = db_connection.execute(
+        "UPDATE Participation SET Score = ? WHERE Id = ?",
+        (score,participation_id))
+    
     #send the request
     db_connection.execute("commit")
 
-    return {"playerName": p.player, "score": score}
+    return_json = {"playerName": p.player, "score": score}
+    
+    return return_json, 200
 
 def RemoveParticipations():
     #création d'un objet connection
